Complete checksum output.

+++ src/util/ChecksumCalculator.cpp
@@ -5,9 +5,9 @@
 #include "ChecksumCalculator.h"
 
 
-ChecksumCalculator::ChecksumCalculator(const unsigned int &num_channels) {
+ChecksumCalculator::ChecksumCalculator(const unsigned long &num_channels) {
     current_index = 0;
-    for (unsigned int c = 0; c < num_channels; c++) {
+    for (unsigned long c = 0; c < num_channels; c++) {
         checksums.push_back(0);
     }
 }
@@ -17,7 +17,7 @@ void ChecksumCalculator::add(__int16_t value) {
     checksums[current_channel] += value;
 }
 
-__int16_t ChecksumCalculator::get(const unsigned int &channel_index) {
+__int16_t ChecksumCalculator::get(const unsigned long &channel_index) {
     return checksums[channel_index];
 }
